% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.gof.R
\name{chi2.gof}
\alias{chi2.gof}
\title{The Cressie-Read test for continuous data.}
\usage{
chi2.gof(x, dist="norm", lambda=1)
}
\arguments{
\item{x}{a numeric vector of values.}

\item{dist}{names distributions: norm (deflaut), gamma, beta and betap (aka Beta Prime).}

\item{lambda}{numeric parametr}
}
\value{
An object of class "htest" containing the following components:
\describe{
\item{estimate}{estimates distribution parameters and loglik}
\item{parameter}{the degrees of freedom and lambda parameter}
\item{statistic}{statistic test value}
\item{p.value}{p-value}
         }
}
\description{
Goodnes of fit test for distributions: normal, gamma, beta and betap (aka Beta Prime).
}
\note{
estimate parameter in \code{\link[fitdistrplus]{fitdist}}.
}
\examples{
set.seed(2305)
g <- rnorm(120)
chi2.gof(g)
}
\references{
Noel Cressie and Timothy R. C. Read (1984).
Multinomial Goodness-of-Fit Test. Journal of the Royal Statistical Society. Series B (Methodological), Vol. 46, No. 3 (1984), 440-464.
}
\seealso{
\code{\link[nortest]{pearson.test}} \code{\link{ks.test}} \code{\link[ADGofTest]{ad.test}} \code{\link[vsgoftest]{vs.test}}
}
\author{
Krzysztof Trajkowski
}
\keyword{chi2.gof}
