% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.test.R
\name{chi2.test}
\alias{chi2.test}
\title{The Cressie-Read test for discrete data.}
\usage{
chi2.test(x, dist = "pois", lambda = 1)
}
\arguments{
\item{x}{a numeric vector of values.}

\item{dist}{names distributions: pois (deflaut) or nbinom}

\item{lambda}{parameter lambda test Cressie-Read.}
}
\value{
An object of class "htest" containing the following components:
\describe{
\item{estimate}{estimates distribution parameters and loglik}
\item{parameter}{the degrees of freedom and lambda parameter}
\item{statistic}{statistic test value}
\item{p.value}{p-value}
         }
}
\description{
Goodnes of fit test for two distributions: Poisson and negative binomial.
}
\examples{
set.seed(2305)
g <- rpois(120,3)
chi2.test(g)
}
\references{
Noel Cressie and Timothy R. C. Read (1984).
Multinomial Goodness-of-Fit Test. Journal of the Royal Statistical Society. Series B (Methodological), Vol. 46, No. 3 (1984), 440-464.
}
\seealso{
\code{\link[vcd]{goodfit}}
}
\author{
Krzysztof Trajkowski
}
\keyword{chi2.test}
