\name{CressieReadTest-package}
\alias{CressieReadTest-package}
\alias{CressieReadTest}
\docType{package}
\title{
The D-squared Cressie-Read test
}
\description{
The D-squared Cressie-Read test
}
\details{
\tabular{ll}{
Package: \tab CressieReadTest\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2013-05-25\cr
License: \tab GPL-2\cr
}
}
\author{
Krzysztof Trajkowski

Maintainer: Krzysztof Trajkowski
<seaproject@poczta.onet.pl>
}
