% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{obs}
\alias{obs}
\title{The observed and expected counts.}
\usage{
obs(x, dist = "pois")
}
\arguments{
\item{x}{a numeric vector of values.}

\item{dist}{names distributions: pois (deflaut), binom or nbinom}
}
\value{
An object containing the following components:
\describe{
\item{ob}{observed counts}
\item{ex}{expected counts}
         }
}
\description{
The observed and expected counts under the null hypothesis: Poisson, binomial and negative binomial.
}
\examples{
set.seed(2305)
g <- rpois(120,3)
obs(g)
}
\references{
Noel Cressie and Timothy R. C. Read (1984).
Multinomial Goodness-of-Fit Test. Journal of the Royal Statistical Society. Series B (Methodological), Vol. 46, No. 3 (1984), 440-464.
}
\seealso{
\code{\link[stats]{chisq.test}}
}
\author{
Krzysztof Trajkowski
}
\keyword{chi2.test}
