% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allcr.test.R
\name{allcr.test}
\alias{allcr.test}
\title{All tests of the family Cressie-Read for contingency table.}
\usage{
allcr.test(x, lambda = 2/3)
}
\arguments{
\item{x}{contingency table.}

\item{lambda}{parameter lambda test Cressie-Read.}
}
\value{
The whole family of tests Cressie-Read and contingency coefficient: Q_Yulea, Phi, C-Pearson, C_adj, V-Cramer, T-Tschuprow for 2x2 table or
Phi, C-Pearson, C_adj, V-Cramer, T-Tschuprow for rxc table where r>2 or c>2
}
\description{
All tests of the family Cressie-Read and contingency coefficient.
}
\examples{
# data:
m <- matrix(c(23,32,45,26),2,2)
# all test:
allcr.test(m)
}
\references{
Noel Cressie and Timothy R. C. Read (1984).
Multinomial Goodness-of-Fit Test. Journal of the Royal Statistical Society. Series B (Methodological), Vol. 46, No. 3 (1984), 440-464.
}
\seealso{
\code{\link[vcd]{goodfit}}
}
\author{
Krzysztof Trajkowski
}
\keyword{allcr.test}
