% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff.R
\name{coeff}
\alias{coeff}
\title{The contingency coefficient}
\usage{
coeff(x)
}
\arguments{
\item{x}{contingency table.}
}
\description{
Computes contingency coefficient: C-Pearson, T-Czupurow, Fi-Yule'a, V-Cramer.
}
\examples{
# data:
m <- matrix(c(23,32,45,26),2,2)
# contingency coefficient:
coeff(m)
}
\references{
Noel Cressie and Timothy R. C. Read (1984).
Multinomial Goodness-of-Fit Test. Journal of the Royal Statistical Society. Series B (Methodological), Vol. 46, No. 3 (1984), 440-464.
}
\seealso{
\code{\link{chisq.test}} \code{\link[vcd]{assocstats}}
}
\author{
Krzysztof Trajkowski
}
\keyword{coeff}
