% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.test.R
\name{cr.test}
\alias{cr.test}
\title{The D-squared Cressie-Read test}
\usage{
cr.test(x, lambda=2/3)
}
\arguments{
\item{x}{contingency table.}

\item{lambda}{lambda parameter lambda test Cressie-Read.}
}
\value{
An object of class "htest" containing the following components:
\describe{
\item{parameter}{the degrees of freedom and lambda parameter}
\item{statistic}{statistic test value}
\item{p.value}{p-value}
         }
}
\description{
Computes D-squared Cressie-Read test.
}
\examples{
# data:
m <- matrix(c(23,32,45,26),2,2)

# Cressie-Read:
cr.test(m)

# Pearson:
cr.test(m,lambda=1)

# Likelihood Ratio:
cr.test(m,lambda=1e-05)

# Freeman-Tukey's:
cr.test(m,lambda=-1/2)

# Neyman's:
cr.test(m,lambda=-2)

# Kullback-Leibler:
cr.test(m,lambda=-0.99999)
}
\references{
Noel Cressie and Timothy R. C. Read (1984).
Multinomial Goodness-of-Fit Test. Journal of the Royal Statistical Society. Series B (Methodological), Vol. 46, No. 3 (1984), 440-464.
}
\seealso{
\code{\link{chisq.test}}   \code{\link{fisher.test}}
}
\author{
Krzysztof Trajkowski
}
\keyword{cr.test}
